pyproject.toml
[tool.poetry]
name = "xfinity-etl"
version = "0.1.0"
description = ""
authors = ["Thomas Dziedzic <thomasdziedzic0@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
selenium = "^3.141.0"
snowflake-connector-python = "^2.3.10"
python-dotenv = "^0.15.0"

[tool.poetry.dev-dependencies]
bpython = "^0.21"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
